catalog-service/requirements.txt
Flask
pymongo
werkzeug 
python-dotenv
gunicorn
pyjwt
Flask-CORS
